app/agents/meta_agents/architecture_design_agent.py
from typing import Any, Dict, List
from app.agents.meta_agents.meta_agents import MetaAgent
from app.tools.reasoning_engine import ReasoningEngine
from app.tools.ontology_reasoner import OntologyReasoner
from app.models.knowledge.knowledge_base import KnowledgeBase
from app.tools.llm_manager import LLMManager
from app.models.knowledge.ontology.ontology import Ontology
import asyncio

class ArchitectureDesignAgent(MetaAgent):
    """
    Designs the overall architecture of the domain-specific MAS.
    
    Key functions:
    - Define the high-level structure of the MAS
    - Determine communication protocols and interaction patterns
    - Design data flow and storage mechanisms
    - Ensure alignment with TOGAF enterprise architecture principles
    """
    def __init__(self, agent_id: str, name: str, api_key: str, llm_service: Any, agent_communication_service: Any):
        super().__init__(
            agent_id=agent_id,
            name=name,
            api_key=api_key,
            llm_service=llm_service,
            agent_communication_service=agent_communication_service
        )
        self.agent_type = "architecture_design"
        self.knowledge_base = KnowledgeBase()
        self.reasoning_engine = ReasoningEngine(self.knowledge_base, api_key)
        self.llm_manager = LLMManager()
        self.ontology = Ontology()  # Initialize with your architecture design ontology
        self.ontology_reasoner = OntologyReasoner(self.llm_manager, self.ontology)

        self._init_beliefs()
        self._init_desires()
        self._init_goals()
        self._init_plans()

    def _init_beliefs(self):
        self.add_belief("Architecture design is crucial for MAS scalability and performance")
        self.add_belief("TOGAF principles guide effective enterprise architecture")

    def _init_desires(self):
        self.add_desire("Design a scalable and efficient MAS architecture", priority=10)
        self.add_desire("Ensure architecture aligns with enterprise goals", priority=9)

    def _init_goals(self):
        self.add_goal("Define high-level MAS structure", priority=9)
        self.add_goal("Determine communication protocols", priority=8)
        self.add_goal("Design data flow mechanisms", priority=8)
        self.add_goal("Ensure TOGAF alignment", priority=7)

    def _init_plans(self):
        self.create_plan(
            self.goals[0].id,
            [
                "Analyze domain requirements",
                "Identify key components",
                "Define component interactions",
                "Create architecture diagram"
            ]
        )

    async def reason(self):
        print("Starting reasoning process for architecture design")
        
        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        updated_beliefs = await self.reasoning_engine.reason({"beliefs": current_beliefs})
        
        for belief in updated_beliefs.get("beliefs", []):
            self.add_belief(belief["content"])

        new_knowledge = await self.ontology_reasoner.infer_new_knowledge()
        
        for concept in new_knowledge.get("new_concepts", []):
            self.add_belief(f"New architecture concept: {concept['name']} - {concept['description']}")
        
        for relationship in new_knowledge.get("new_relationships", []):
            self.add_belief(f"New architecture relationship: {relationship['name']} between {relationship['domain']} and {relationship['range']}")

        architecture_query = "What are the key architectural patterns suitable for this MAS?"
        architecture_patterns = await self.ontology_reasoner.answer_query(architecture_query)
        self.add_belief(f"Suitable architectural patterns: {architecture_patterns}")

        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        new_desires = await self.reasoning_engine.generate_desires(current_beliefs)
        
        for desire in new_desires:
            self.add_desire(desire["description"], desire["priority"])

        print("Reasoning process for architecture design completed")

    async def plan(self):
        print("Starting planning process for architecture design")
        for goal in self.goals:
            if goal.description == "Define high-level MAS structure":
                self.create_plan(
                    goal.id,
                    [
                        "Identify key components",
                        "Define component interactions",
                        "Create architecture diagram",
                        "Validate architecture with stakeholders"
                    ]
                )
        
        current_state = self.get_current_state()
        optimized_plan = await self.reasoning_engine.reason_and_plan(self.goals[0].description, current_state)
        
        if optimized_plan.get("plan"):
            self.update_plan(self.goals[0].id, optimized_plan["plan"].steps)
        
        print("Planning process for architecture design completed")

    async def execute(self):
        print("Starting execution process for architecture design")
        for plan in self.plans:
            for task in plan.steps:
                if task.status == "pending":
                    print(f"Executing task: {task.description}")
                    try:
                        execution_result = await self.reasoning_engine.simulate_action(task.description, self.get_current_state())
                        self.update_task_status(task.id, "completed")
                        
                        for key, value in execution_result.items():
                            self.add_belief(f"Architecture design task result - {key}: {value}")
                    except Exception as e:
                        print(f"Error executing architecture design task {task.description}: {str(e)}")
                        self.update_task_status(task.id, "failed")
        print("Execution process for architecture design completed")

    def get_current_state(self) -> Dict[str, Any]:
        return {
            "beliefs": [belief.to_dict() for belief in self.beliefs],
            "desires": [desire.to_dict() for desire in self.desires],
            "goals": [goal.to_dict() for goal in self.goals],
            "plans": [plan.to_dict() for plan in self.plans],
        }

    async def run(self):
        while True:
            await self.reason()
            await self.plan()
            await self.execute()
            await asyncio.sleep(1)  # Adjust the sleep time as needed

    async def process(self, input_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Design the overall architecture of the domain-specific MAS based on input data.

        Args:
            input_data (Dict[str, Any]): The input data required for designing the MAS architecture.

        Returns:
            Dict[str, Any]: A dictionary containing the designed MAS architecture components.
        """
        await self.reason()  # Update beliefs and goals based on input data
        await self.plan()    # Create or update plans for architecture design

        mas_structure = await self._define_mas_structure(input_data)
        communication_protocols = await self._determine_communication_protocols(mas_structure)
        interaction_patterns = await self._determine_interaction_patterns(mas_structure)
        data_flow_design = await self._design_data_flow(mas_structure)
        storage_mechanisms = await self._design_storage_mechanisms(data_flow_design)
        await self._ensure_togaf_alignment(mas_structure, communication_protocols, interaction_patterns, data_flow_design, storage_mechanisms)
        
        return {
            "mas_structure": mas_structure,
            "communication_protocols": communication_protocols,
            "interaction_patterns": interaction_patterns,
            "data_flow_design": data_flow_design,
            "storage_mechanisms": storage_mechanisms
        }

    async def _define_mas_structure(self, input_data: Dict[str, Any]) -> Dict[str, Any]:
        structure_query = f"Based on this input data: {input_data}, define a high-level MAS structure."
        mas_structure = await self.ontology_reasoner.answer_query(structure_query)
        self.add_belief(f"Defined MAS structure: {mas_structure}")
        return mas_structure

    async def _determine_communication_protocols(self, mas_structure: Dict[str, Any]) -> List[str]:
        protocol_query = f"For this MAS structure: {mas_structure}, determine suitable communication protocols."
        communication_protocols = await self.ontology_reasoner.answer_query(protocol_query)
        self.add_belief(f"Determined communication protocols: {communication_protocols}")
        return communication_protocols

    async def _determine_interaction_patterns(self, mas_structure: Dict[str, Any]) -> List[str]:
        pattern_query = f"For this MAS structure: {mas_structure}, determine appropriate interaction patterns."
        interaction_patterns = await self.ontology_reasoner.answer_query(pattern_query)
        self.add_belief(f"Determined interaction patterns: {interaction_patterns}")
        return interaction_patterns

    async def _design_data_flow(self, mas_structure: Dict[str, Any]) -> Dict[str, Any]:
        data_flow_query = f"For this MAS structure: {mas_structure}, design data flow mechanisms."
        data_flow_design = await self.ontology_reasoner.answer_query(data_flow_query)
        self.add_belief(f"Designed data flow: {data_flow_design}")
        return data_flow_design

    async def _design_storage_mechanisms(self, data_flow_design: Dict[str, Any]) -> Dict[str, Any]:
        storage_query = f"Based on this data flow design: {data_flow_design}, design appropriate storage mechanisms."
        storage_mechanisms = await self.ontology_reasoner.answer_query(storage_query)
        self.add_belief(f"Designed storage mechanisms: {storage_mechanisms}")
        return storage_mechanisms

    async def _ensure_togaf_alignment(self, mas_structure: Dict[str, Any], communication_protocols: List[str],
                                interaction_patterns: List[str], data_flow_design: Dict[str, Any],
                                storage_mechanisms: Dict[str, Any]) -> None:
        alignment_query = f"""
        Ensure TOGAF alignment for:
        MAS structure: {mas_structure}
        Communication protocols: {communication_protocols}
        Interaction patterns: {interaction_patterns}
        Data flow design: {data_flow_design}
        Storage mechanisms: {storage_mechanisms}
        """
        alignment_result = await self.ontology_reasoner.answer_query(alignment_query)
        self.add_belief(f"TOGAF alignment result: {alignment_result}")
